pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-DesignGym"
version = "0.1.0"
description = "Designgym environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
  "openenv-core>=0.2.2",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "openai>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m DesignGym.server.app
server = "DesignGym.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["DesignGym", "DesignGym.server"]
package-dir = { "DesignGym" = ".", "DesignGym.server" = "server" }

